Factor out RemoteModRegistry
